V1 of my one tap quest bot
V1 Of my one tap quest bot V1 uses pyautogui to find and detect a certain pixel color on the mug. Unfortunately, before starting this project I did not realize that the hilt of the sword in this game is the same color as the mug, so it only is optimized 50% of the time. In V2 i will use OpenCV to detect the mug, rather than just the pixel color. V2 coming soon.